pyproject.toml
[tool.poetry]
name = "airbnbclone-study"
version = "0.1.0"
description = ""
authors = ["bioroid17 <trace1997@ajou.ac.kr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
